requirements-dev.txt
# FirmXtract — Development dependencies
# Install with: pip install -r requirements-dev.txt
#
# These mirror [project.optional-dependencies.dev] in pyproject.toml.
# Keep both files in sync when updating versions.
#
# For runtime deps only: pip install -e .
# For everything:        pip install -e ".[dev]"

# ── Testing ──────────────────────────────────────────────────────────────────
pytest>=8.0.0,<9.0
pytest-cov>=5.0.0,<6.0
pytest-mock>=3.14.0,<4.0

# ── Linting & Formatting ─────────────────────────────────────────────────────
ruff>=0.4.0,<1.0
black>=24.0.0,<25.0

# ── Type Checking ─────────────────────────────────────────────────────────────
mypy>=1.10.0,<2.0
pyserial-stubs>=0.1        # type stubs for pyserial

# ── Runtime (duplicated here so `pip install -r` works standalone) ────────────
typer[all]>=0.12.0,<1.0
pyserial>=3.5,<4.0
rich>=13.0.0,<14.0
